requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
python-dotenv==1.0.1
scipy==1.12.0
